+++ ecdsa.h
@@ -71,9 +71,9 @@ int ecdsa_sign_digest(const ecdsa_curve *curve, const uint8_t *priv_key, const u
 void ecdsa_get_public_key33(const ecdsa_curve *curve, const uint8_t *priv_key, uint8_t *pub_key);
 void ecdsa_get_public_key65(const ecdsa_curve *curve, const uint8_t *priv_key, uint8_t *pub_key);
 void ecdsa_get_pubkeyhash(const uint8_t *pub_key, uint8_t *pubkeyhash);
-void ecdsa_get_address_raw(const uint8_t *pub_key, uint8_t version, uint8_t *addr_raw);
-void ecdsa_get_address(const uint8_t *pub_key, uint8_t version, char *addr, int addrsize);
-void ecdsa_get_wif(const uint8_t *priv_key, uint8_t version, char *wif, int wifsize);
+void ecdsa_get_address_raw(const uint8_t *pub_key, uint32_t version, uint8_t *addr_raw);
+void ecdsa_get_address(const uint8_t *pub_key, uint32_t version, char *addr, int addrsize);
+void ecdsa_get_wif(const uint8_t *priv_key, uint32_t version, char *wif, int wifsize);
 
 int ecdsa_address_decode(const char *addr, uint8_t *out);
 int ecdsa_read_pubkey(const ecdsa_curve *curve, const uint8_t *pub_key, curve_point *pub);
